chore: move test to appropriate place
